♻️ refactor(api): migrate password and session-token services to application layer
