+++ apps/api/src/app/auth/[client]/logout.ts
@@ -1,4 +1,5 @@
 import { t } from "elysia";
+import { SessionTokenService } from "../../../application/services/session-token";
 import { LogoutUseCase } from "../../../application/use-cases/auth";
 import { SESSION_COOKIE_NAME } from "../../../common/constants";
 import { clientSchema } from "../../../common/schema";
@@ -7,7 +8,6 @@ import { SessionRepository } from "../../../interface-adapter/repositories/sessi
 import { authGuard } from "../../../modules/auth-guard";
 import { CookieService } from "../../../modules/cookie";
 import { ElysiaWithEnv } from "../../../modules/elysia-with-env";
-import { SessionTokenService } from "../../../services/session-token";
 
 const cookieSchemaObject = {
 	[SESSION_COOKIE_NAME]: t.Optional(t.String()),
