+++ apps/api/src/app/auth/[client]/signup/index.ts
@@ -1,4 +1,6 @@
 import { t } from "elysia";
+import { PasswordService } from "../../../../application/services/password";
+import { SessionTokenService } from "../../../../application/services/session-token";
 import { SignupUseCase } from "../../../../application/use-cases/auth";
 import { SESSION_COOKIE_NAME } from "../../../../common/constants";
 import { clientSchema } from "../../../../common/schema";
@@ -9,8 +11,6 @@ import { UserCredentialRepository } from "../../../../interface-adapter/reposito
 import { CookieService } from "../../../../modules/cookie";
 import { ElysiaWithEnv } from "../../../../modules/elysia-with-env";
 import { rateLimiter } from "../../../../modules/rate-limiter";
-import { PasswordService } from "../../../../services/password";
-import { SessionTokenService } from "../../../../services/session-token";
 import { Provider } from "./[provider]";
 
 const cookieSchemaObject = {
