+++ apps/api/src/application/services/password/index.ts
