+++ apps/api/src/application/services/password/interfaces/password.service.interface.ts
