+++ apps/api/src/application/services/session-token/interfaces/session-token.service.interface.ts
