+++ apps/api/src/application/use-cases/auth/login.usecase.ts
@@ -2,8 +2,8 @@ import { sessionExpiresSpan } from "../../../common/constants";
 import type { ISessionRepository } from "../../../interface-adapter/repositories/session";
 import type { IUserRepository } from "../../../interface-adapter/repositories/user";
 import type { IUserCredentialRepository } from "../../../interface-adapter/repositories/user-credential";
-import type { IPasswordService } from "../../../services/password";
-import type { ISessionTokenService } from "../../../services/session-token";
+import type { IPasswordService } from "../../services/password";
+import type { ISessionTokenService } from "../../services/session-token";
 import type { ILoginUseCase, ILoginUseCaseResult } from "./interfaces/login.usecase.interface";
 
 export class LoginUseCase implements ILoginUseCase {
